Cleanup header, pull out search into its own component
