+++ src/pages/type-charts.js
@@ -7,12 +7,7 @@ import Header from "../components/MainHeader"
 const TypeCharts = () => {
   return (
     <Layout>
-      <Header
-        showArrow={true}
-        heading="Type Charts"
-        pokemonList={[]}
-        setSearchResults={() => {}} // don't do anything right now
-      />
+      <Header heading="Type Charts" />
       <p sx={{ textAlign: "center" }}>Page under construction</p>
     </Layout>
   )
