+++ src/templates/home.js
@@ -5,19 +5,20 @@ import { useState } from "react"
 import Layout from "../components/Layout"
 import Header from "../components/MainHeader"
 import PokemonList from "../components/PokemonList"
+import { Search } from "../components/SearchNew"
 
 const Home = ({ pageContext: { allPokemon } }) => {
   const [searchResults, setSearchResults] = useState([...allPokemon])
 
   return (
     <Layout>
-      <Header
-        showArrow={true}
-        heading="Pokédex"
-        setSearchResults={setSearchResults}
-        searchResults={searchResults}
-        pokemonList={allPokemon}
-      />
+      <Header heading="Pokédex">
+        <Search
+          setSearchResults={setSearchResults}
+          searchResults={searchResults}
+          pokemonList={allPokemon}
+        />
+      </Header>
       <PokemonList pokemonList={searchResults} />
     </Layout>
   )
